[WB-6939] Reorg wandb/sdk/interface (prep for fork support) (#2816)
