+++ mypy.ini
@@ -72,7 +72,40 @@ check_untyped_defs = True
 disallow_untyped_decorators = True
 strict_equality = True
 
+[mypy-wandb.sdk.interface.interface_queue]
+disallow_incomplete_defs = True
+disallow_untyped_defs = True
+warn_unused_ignores = True
+warn_return_any = True
+warn_unreachable = True
+check_untyped_defs = True
+disallow_untyped_calls = True
+disallow_untyped_decorators = True
+strict_equality = True
+
+[mypy-wandb.sdk.interface.interface_grpc]
+disallow_incomplete_defs = True
+disallow_untyped_defs = True
+warn_unused_ignores = True
+warn_return_any = True
+warn_unreachable = True
+check_untyped_defs = True
+disallow_untyped_calls = True
+disallow_untyped_decorators = True
+strict_equality = True
+
 [mypy-wandb.sdk.interface.message_future]
+disallow_incomplete_defs = True
+disallow_untyped_defs = True
+warn_unused_ignores = True
+warn_return_any = True
+warn_unreachable = True
+check_untyped_defs = True
+disallow_untyped_calls = True
+disallow_untyped_decorators = True
+strict_equality = True
+
+[mypy-wandb.sdk.interface.message_future_poll]
 disallow_incomplete_defs = True
 disallow_untyped_defs = True
 warn_unused_ignores = True
@@ -326,17 +359,6 @@ disallow_untyped_decorators = True
 strict_equality = True
 
 [mypy-wandb.sdk.backend.backend]
-disallow_incomplete_defs = True
-disallow_untyped_defs = True
-warn_unused_ignores = True
-warn_return_any = True
-warn_unreachable = True
-check_untyped_defs = True
-disallow_untyped_calls = True
-disallow_untyped_decorators = True
-strict_equality = True
-
-[mypy-wandb.sdk.interface.iface_grpc]
 disallow_incomplete_defs = True
 disallow_untyped_defs = True
 warn_unused_ignores = True
