+++ wandb/sdk/internal/stats.py
@@ -14,7 +14,7 @@ from wandb import util
 from wandb.vendor.pynvml import pynvml
 
 from . import tpu
-from ..interface.interface import BackendSender
+from ..interface.interface_queue import InterfaceQueue
 from ..lib import telemetry
 
 
@@ -66,13 +66,13 @@ def gpu_in_use_by_this_process(gpu_handle: GPUHandle) -> bool:
 class SystemStats(object):
 
     _pid: int
-    _interface: BackendSender
+    _interface: InterfaceQueue
     sampler: SamplerDict
     samples: int
     _thread: Optional[threading.Thread]
     gpu_count: int
 
-    def __init__(self, pid: int, interface: BackendSender) -> None:
+    def __init__(self, pid: int, interface: InterfaceQueue) -> None:
         try:
             pynvml.nvmlInit()
             self.gpu_count = pynvml.nvmlDeviceGetCount()
